refactor map screen
